requirements.txt
langchain==0.3.20
langchain-community==0.3.18
langchain-huggingface==0.1.2
huggingface-hub==0.28.1
streamlit==1.42.0
python-dotenv==1.0.0
